% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNBP.R
\name{BNBP}
\alias{BNBP}
\title{Beta-Negative Binomial process modeling of sequencing count data}
\usage{
BNBP(y, groups, randtry = 1, Burnin = 1000L, Collections = 1000L)
}
\arguments{
\item{y}{a matrix of counts, rows are corresponding to genes and columns are
corresponding to samples.}

\item{groups}{a vector of factors indicating group memberships of samples}

\item{randtry}{state of set.seed}

\item{Burnin}{Number of burn-in iterations in MCMC}

\item{Collections}{Number of collected posterior samples after burn-in}
}
\value{
posterior samples of probability and dispersion parameters of Negative Binomial
}
\description{
This function fits BNBP model to counts in different conditions and returns the
posterior samples of probability and dispersion parameters than can be used to
assess the differential expression significance.
}
\examples{
y <- rnbinom(n = 100, size = 5, prob = 0.5)
y <- matrix(data = y, nrow = 10, ncol = 10)
groups <- factor(rep(c(1,2), each = 5))
res <- BNBP(y, groups, randtry = 1, Burnin = 1000L, Collections = 1000L)
p <- res$p_k
p <- p/(1-p)
gene.kl <- KLsym(p[,1:1000], p[,1001:2000])
}
