% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNBP.R
\name{GNBP}
\alias{GNBP}
\title{Gamma-Negative Binomial process modeling of sequencing count data}
\usage{
GNBP(y, groups, randtry = 1, Burnin = 1000L, Collections = 1000L)
}
\arguments{
\item{y}{a matrix of counts, rows are corresponding to genes and columns are
corresponding to samples.}

\item{groups}{a vector of factors indicating group memberships of samples}

\item{randtry}{state of set.seed}

\item{Burnin}{Number of burn-in iterations in MCMC}

\item{Collections}{Number of collected posterior samples after burn-in}
}
\value{
posterior samples of probability and dispersion parameters of Negative Binomial
}
\description{
This function fits GNBP model to counts in different conditions and returns the
posterior samples of probability and dispersion parameters than can be used to
assess the differential expression significance.
}
\examples{
y <- rnbinom(n = 100, size = 5, prob = 0.5)
y <- matrix(data = y, nrow = 10, ncol = 10)
groups <- factor(rep(c(1,2), each = 5))
res <- GNBP(y, groups, randtry = 1, Burnin = 1000L, Collections = 1000L)
r <- res$r_k
gene.kl <- KLsym(r[,1:1000], r[,1001:2000])
}
