% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLsym.R
\name{KLsym}
\alias{KLsym}
\title{Symmetric Kullback-Leibler divergence between samples drawn from random variables}
\usage{
KLsym(x1, x2)
}
\arguments{
\item{x1}{a matrix, where each row corresponds to draws from a random distribution}

\item{x2}{a matrix, where each row corresponds to draws from a random distribution
and should have the same number of rows as \code{x1}}
}
\value{
A vector with the length equal to the number of rows of input matrices,
}
\description{
This function calculates the symmetric Kullback-Leibler divergence between two
empirical distriburions.
}
\examples{
x1 <- matrix(0, 2, 100)
x1[1,] <- rnorm(100, 0, 1)
x1[2,] <- rnorm(100, -1, 2)
x2 <- matrix(0, 2, 100)
x2[1,] <- rnorm(100, 1, 0)
x2[1,] <- rnorm(100, 1, 3)
kl <- KLsym(x1, x2)
}
